testing/web-platform/mozilla/tests/baselinecoverage/wpt_baselinecoverage.html

<!doctype html>
<meta charset=utf-8>
<title>Baseline Coverage</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
  test(function() {
    assert_equals(true, true);
  }, "Collecting baseline coverage for web-platform tests.");
</script>
